WIP move About page to refactor
