+++ wv2util/MainWindow.xaml
@@ -9,15 +9,6 @@
         Icon="logo.ico"
         Title="WebView2Utilities"
         Height="700" Width="1100" ResizeMode="CanResizeWithGrip" MinWidth="662" MinHeight="360">
-    <Window.Resources>
-        <Style x:Key="heading" TargetType="Label">
-            <Setter Property="FontSize" Value="24" />
-        </Style>
-
-        <Style x:Key="subHeading" TargetType="Label">
-            <Setter Property="FontSize" Value="16" />
-        </Style>
-    </Window.Resources>
     <Grid>
         <Grid.ColumnDefinitions>
             <ColumnDefinition/>
@@ -90,37 +81,12 @@
                         <TextBlock Text="About"/>
                     </StackPanel>
                 </TabItem.Header>
-                <TextBlock HorizontalAlignment="Stretch" VerticalAlignment="Stretch" Margin="20" >
-          <Image Source="logo.ico" Height="128" Width="128" Margin="0,0,20,0"/>
-          <TextBlock>
-            <Label Style="{StaticResource heading}" Content="WebView2Utilities" Margin="0,0,0,0" HorizontalAlignment="Left"/>
-            <LineBreak/>
-            <TextBlock FontWeight="Bold" x:Name="VersionInfo">vN.N.N.N</TextBlock>
-            <LineBreak/>
-            <TextBlock>This is WebView2Utilities. It helps with developing and debugging Microsoft Edge WebView2 usage.</TextBlock>
-            <LineBreak/>
-            <Hyperlink NavigateUri="https://github.com/david-risney/WebView2Utilities" RequestNavigate="Hyperlink_RequestNavigate">
-              WebView2 Tools GitHub
-            </Hyperlink>
-            <LineBreak/>
-            <Hyperlink NavigateUri="https://docs.microsoft.com/en-us/microsoft-edge/webview2/" RequestNavigate="Hyperlink_RequestNavigate">
-              WebView2 Documentation
-            </Hyperlink>
-            <LineBreak/>
-            <LineBreak/>
-          </TextBlock>
-          <LineBreak/>
-          <TextBlock FontSize="16" Margin="0,10,0,10">Latest WebView2 News</TextBlock>
-          <LineBreak/>
-          <InlineUIContainer>
-          <StackPanel Margin="10,0,0,0" x:Name="NewsPanel" />
-          </InlineUIContainer>
-          <LineBreak/>
-          <Hyperlink NavigateUri="https://github.com/MicrosoftEdge/WebView2Announcements/issues" RequestNavigate="Hyperlink_RequestNavigate">
-              More news ...
-          </Hyperlink>
-        </TextBlock>
-      </TabItem>
+                <Frame>
+                    <Frame.Content>
+                        <local:AboutPage x:Name="AboutPage"/>
+                    </Frame.Content>
+                </Frame>
+            </TabItem>
         </TabControl>
     </Grid>
 </Window>
