+++ wv2util/Pages/AboutPage.xaml
@@ -1,0 +1,49 @@
+﻿<Page x:Class="wv2util.AboutPage"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+      xmlns:local="clr-namespace:wv2util"
+      mc:Ignorable="d" 
+      d:DesignHeight="450" d:DesignWidth="800"
+      Title="AboutPage">
+    <Page.Resources>
+        <Style x:Key="heading" TargetType="Label">
+            <Setter Property="FontSize" Value="24" />
+        </Style>
+
+        <Style x:Key="subHeading" TargetType="Label">
+            <Setter Property="FontSize" Value="16" />
+        </Style>
+    </Page.Resources>
+    <TextBlock HorizontalAlignment="Stretch" VerticalAlignment="Stretch" Margin="20" >
+        <Image Source="../logo.ico" Height="128" Width="128" Margin="0,0,20,0"/>
+        <TextBlock>
+            <Label Style="{StaticResource heading}" Content="WebView2Utilities" Margin="0,0,0,0" HorizontalAlignment="Left"/>
+            <LineBreak/>
+            <TextBlock FontWeight="Bold" x:Name="VersionInfo">vN.N.N.N</TextBlock>
+            <LineBreak/>
+            <TextBlock>This is WebView2Utilities. It helps with developing and debugging Microsoft Edge WebView2 usage.</TextBlock>
+            <LineBreak/>
+            <Hyperlink NavigateUri="https://github.com/david-risney/WebView2Utilities" RequestNavigate="Hyperlink_RequestNavigate">
+              WebView2 Tools GitHub
+            </Hyperlink>
+            <LineBreak/>
+            <Hyperlink NavigateUri="https://docs.microsoft.com/en-us/microsoft-edge/webview2/" RequestNavigate="Hyperlink_RequestNavigate">
+              WebView2 Documentation
+            </Hyperlink>
+            <LineBreak/>
+            <LineBreak/>
+        </TextBlock>
+        <LineBreak/>
+        <TextBlock FontSize="16" Margin="0,10,0,10">Latest WebView2 News</TextBlock>
+        <LineBreak/>
+        <InlineUIContainer>
+            <StackPanel Margin="10,0,0,0" x:Name="NewsPanel" />
+        </InlineUIContainer>
+        <LineBreak/>
+        <Hyperlink NavigateUri="https://github.com/MicrosoftEdge/WebView2Announcements/issues" RequestNavigate="Hyperlink_RequestNavigate">
+              More news ...
+        </Hyperlink>
+    </TextBlock>
+</Page>
